+++ ent/metadataformat/where.go
@@ -9,28 +9,28 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.MetadataFormat {
+func ID(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.MetadataFormat {
+func IDEQ(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.MetadataFormat {
+func IDNEQ(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldID), id))
 	})
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.MetadataFormat {
+func IDIn(ids ...int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		v := make([]any, len(ids))
 		for i := range v {
@@ -41,7 +41,7 @@ func IDIn(ids ...int) predicate.MetadataFormat {
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.MetadataFormat {
+func IDNotIn(ids ...int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		v := make([]any, len(ids))
 		for i := range v {
@@ -52,28 +52,28 @@ func IDNotIn(ids ...int) predicate.MetadataFormat {
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.MetadataFormat {
+func IDGT(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldID), id))
 	})
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.MetadataFormat {
+func IDGTE(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldID), id))
 	})
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.MetadataFormat {
+func IDLT(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldID), id))
 	})
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.MetadataFormat {
+func IDLTE(id int64) predicate.MetadataFormat {
 	return predicate.MetadataFormat(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldID), id))
 	})
